requirements.txt
pyseneye
paho-mqtt
